feat: navBar - 작업 중
