+++ src/global.d.ts
@@ -1,0 +1,4 @@
+declare module "*.css" {
+  const content: { [className: string]: string };
+  export default content;
+}
