+++ src/style/layout.css
@@ -1,0 +1,22 @@
+.app {
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
+.top {
+    flex: 0;
+    border-bottom: solid 1px var(--adm-color-border);
+}
+
+.body {
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.bottom {
+    flex: 0;
+    border-top: solid 1px var(--adm-color-border);
+}
